updated final version

+++ delete_all.py
@@ -4,10 +4,11 @@ import boto3
 s3 = boto3.resource('s3')
 
 for bucket in s3.buckets.all():
-    bucket.objects.all().delete()
-    bucket.delete()
+    if bucket.name != 'jbloggs0pn73a':
+        bucket.objects.all().delete()
+        bucket.delete()
 
-print("All S3 buckets emptied and deleted")
+print("All S3 buckets except jbloggs0pn73a emptied and deleted")
 
 ec2 = boto3.resource('ec2')
